+++ src/ESP8266_PWM_ISR.hpp
@@ -16,7 +16,7 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.2
+  Version: 1.2.3
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -25,6 +25,7 @@
   1.2.0   K Hoang      29/01/2022 Fix multiple-definitions linker error. Improve accuracy
   1.2.1   K Hoang      30/01/2022 Fix bug. Optimize code
   1.2.2   K Hoang      30/01/2022 DutyCycle to be updated at the end current PWM period
+  1.2.3   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period. Optimize code
 *****************************************************************************************************************************/
 
 #pragma once
@@ -37,13 +38,13 @@
 #endif
 
 #ifndef ESP8266_PWM_VERSION
-  #define ESP8266_PWM_VERSION           "ESP8266_PWM v1.2.0"
+  #define ESP8266_PWM_VERSION           "ESP8266_PWM v1.2.3"
   
   #define ESP8266_PWM_VERSION_MAJOR     1
   #define ESP8266_PWM_VERSION_MINOR     2
-  #define ESP8266_PWM_VERSION_PATCH     0
+  #define ESP8266_PWM_VERSION_PATCH     3
 
-  #define ESP8266_PWM_VERSION_INT       1002000
+  #define ESP8266_PWM_VERSION_INT       1002003
 #endif
 
 #ifndef TIMER_INTERRUPT_DEBUG
@@ -112,19 +113,19 @@ class ESP8266_PWM_ISR
     //////////////////////////////////////////////////////////////////
     // PWM
     // Return the channelNum if OK, -1 if error
-    int setPWM(const uint32_t& pin, const double& frequency, const double& dutycycle, esp8266_timer_callback StartCallback = nullptr, 
+    int setPWM(const uint32_t& pin, const float& frequency, const float& dutycycle, esp8266_timer_callback StartCallback = nullptr, 
                 esp8266_timer_callback StopCallback = nullptr)
     {
-      double period = 0.0;
+      uint32_t period = 0;
       
       if ( ( frequency > 0.0 ) && ( frequency <= 500.0 ) )
       {
 #if USING_MICROS_RESOLUTION
       // period in us
-      period = 1000000.0f / frequency;
+      period = (uint32_t) (1000000.0f / frequency);
 #else    
       // period in ms
-      period = 1000.0f / frequency;
+      period = (uint32_t) (1000.0f / frequency);
 #endif
       }
       else
@@ -139,7 +140,7 @@ class ESP8266_PWM_ISR
 
     // period in us
     // Return the channelNum if OK, -1 if error
-    int setPWM_Period(const uint32_t& pin, const double& period, const double& dutycycle, 
+    int setPWM_Period(const uint32_t& pin, const uint32_t& period, const float& dutycycle, 
                       esp8266_timer_callback StartCallback = nullptr, esp8266_timer_callback StopCallback = nullptr)  
     {     
       return setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);      
@@ -149,18 +150,18 @@ class ESP8266_PWM_ISR
     
     // low level function to modify a PWM channel
     // returns the true on success or false on failure
-    bool modifyPWMChannel(const uint8_t& channelNum, const uint32_t& pin, const double& frequency, const double& dutycycle)
+    bool modifyPWMChannel(const uint8_t& channelNum, const uint32_t& pin, const float& frequency, const float& dutycycle)
     {
-      double period = 0.0;
+      uint32_t period = 0;
       
       if ( ( frequency > 0.0 ) && ( frequency <= 500.0 ) )
       {
 #if USING_MICROS_RESOLUTION
       // period in us
-      period = 1000000.0f / frequency;
+      period = (uint32_t) (1000000.0f / frequency);
 #else    
       // period in ms
-      period = 1000.0f / frequency;
+      period = (uint32_t) (1000.0f / frequency);
 #endif
       }
       else
@@ -173,7 +174,7 @@ class ESP8266_PWM_ISR
     }
     
     //period in us
-    bool modifyPWMChannel_Period(const uint8_t& channelNum, const uint32_t& pin, const double& period, const double& dutycycle);
+    bool modifyPWMChannel_Period(const uint8_t& channelNum, const uint32_t& pin, const uint32_t& period, const float& dutycycle);
 
     // destroy the specified PWM channel
     void deleteChannel(const uint8_t& channelNum);
@@ -216,7 +217,7 @@ class ESP8266_PWM_ISR
     // low level function to initialize and enable a new PWM channel
     // returns the PWM channel number (channelNum) on success or
     // -1 on failure (f == NULL) or no free PWM channels 
-    int setupPWMChannel(const uint32_t& pin, const double& period, const double& dutycycle, void* cbStartFunc = nullptr, void* cbStopFunc = nullptr);
+    int setupPWMChannel(const uint32_t& pin, const uint32_t& period, const float& dutycycle, void* cbStartFunc = nullptr, void* cbStopFunc = nullptr);
     
 
     // find the first available slot
@@ -230,7 +231,7 @@ class ESP8266_PWM_ISR
       ///////////////////////////////////
       
       uint64_t      prevTime;           // value returned by the micros() or millis() function in the previous run() call
-      double        period;             // period value, in us / ms
+      uint32_t      period;             // period value, in us / ms
       uint32_t      onTime;             // onTime value, ( period * dutyCycle / 100 ) us  / ms
       
       void*         callbackStart;      // pointer to the callback function when PWM pulse starts (HIGH)
@@ -245,8 +246,9 @@ class ESP8266_PWM_ISR
       bool          enabled;            // true if enabled
       
       // New from v1.3.0     
-      double        newPeriod;          // period value, in us / ms
-      double        newDutyCycle;       // from 0.00 to 100.00, double precision
+      uint32_t      newPeriod;          // period value, in us / ms
+      uint32_t      newOnTime;          // onTime value, ( period * dutyCycle / 100 ) us  / ms
+      float         newDutyCycle;       // from 0.00 to 100.00, float precision
       //////
     } PWM_t;